refactor billingFields component
